Exceptions

+++ src/main/java/com/alkemy/disney/disney/service/impl/CharacterServiceImpl.java
@@ -4,7 +4,7 @@ import com.alkemy.disney.disney.dto.CharacterBasicDTO;
 import com.alkemy.disney.disney.dto.CharacterDTO;
 import com.alkemy.disney.disney.dto.CharacterFiltersDTO;
 import com.alkemy.disney.disney.entity.CharacterEntity;
-import com.alkemy.disney.disney.error.ServiceError;
+import com.alkemy.disney.disney.exception.ParamNotFound;
 import com.alkemy.disney.disney.mapper.CharacterMapper;
 import com.alkemy.disney.disney.repository.CharacterRepository;
 import com.alkemy.disney.disney.repository.specifications.CharacterSpecification;
@@ -29,7 +29,7 @@ public class CharacterServiceImpl implements CharacterService {
     private CharacterMapper characterMapper;
 
     @Transactional
-    public CharacterDTO save(CharacterDTO dto) throws ServiceError {
+    public CharacterDTO save(CharacterDTO dto) throws ParamNotFound {
         CharacterEntity entity = characterMapper.characterDTO2Entity(dto);
         CharacterEntity entitySaved = characterRepository.save(entity);
         CharacterDTO resultDTO = characterMapper.characterEntity2DTO(entitySaved, true);
@@ -37,7 +37,7 @@ public class CharacterServiceImpl implements CharacterService {
     }
 
     @Transactional
-    public CharacterDTO update(Long id, CharacterDTO dto) throws ServiceError {
+    public CharacterDTO update(Long id, CharacterDTO dto) throws ParamNotFound {
         Optional<CharacterEntity> result = characterRepository.findById(id);
         if(result.isPresent()){
             CharacterEntity entity = result.get();
@@ -47,7 +47,7 @@ public class CharacterServiceImpl implements CharacterService {
             CharacterDTO resultDTO = characterMapper.characterEntity2DTO(entitySaved, true);
             return resultDTO;
         }else{
-            throw new ServiceError("id not found");
+            throw new ParamNotFound("character id not found");
         }
     }
 
@@ -57,7 +57,12 @@ public class CharacterServiceImpl implements CharacterService {
     }
 
     public void delete(Long id) {
-        characterRepository.deleteById(id);
+        Optional<CharacterEntity> result = characterRepository.findById(id);
+        if(result.isPresent()){
+            characterRepository.deleteById(id);
+        }else{
+            throw new ParamNotFound("character id not found");
+        }
     }
 
     public List<CharacterBasicDTO> getByFilters(String name, String age, Set<Long> movies, String order){
